@@ -19,6 +19,11 @@ using System.Security.Claims;
 
 namespace SjaInNumbers.Server.Controllers;
 
+/// <summary>
+/// Controller for management of hours.
+/// </summary>
+/// <param name="hoursService">The service for managing hours.</param>
+/// <param name="logger">The logger for this instance.</param>
 [ApiController]
 [Route("api/hours")]
 public class HoursController(IHoursService hoursService, ILogger<HoursController> logger) : ControllerBase
@@ -26,11 +31,19 @@ public class HoursController(IHoursService hoursService, ILogger<HoursController
     private readonly ILogger logger = logger;
     private readonly IHoursService hoursService = hoursService;
 
+    /// <summary>
+    /// Gets the current NHSE Target.
+    /// </summary>
+    /// <param name="etag">The Etag for the data currently held by the client.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the asynchronous operation. Resolves to the result of the action.
+    /// </returns>
     [HttpGet("target")]
     [ProducesResponseType<HoursTarget>(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status304NotModified)]
     [RevalidateCache]
     [Authorize(Policy = "Approved")]
-    public async Task<IActionResult> GetTargetAsync([FromHeader(Name = "If-None-Match")] string? etag)
+    public async Task<ActionResult<HoursTarget>> GetTargetAsync([FromHeader(Name = "If-None-Match")] string? etag)
     {
         var target = await hoursService.GetNhseTargetAsync();
         var actualEtagValue = await hoursService.GetNhseTargetEtagAsync();
@@ -53,13 +66,23 @@ public class HoursController(IHoursService hoursService, ILogger<HoursController
         });
     }
 
+    /// <summary>
+    /// Gets the count of hours around a given date.
+    /// </summary>
+    /// <param name="etag">The Etag for the data currently held by the client.</param>
+    /// <param name="date">The reporting date to be based on.</param>
+    /// <param name="dateType">The date type to use for the report.</param>
+    /// <param name="future">Should only future information be included.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the asynchronous operation. Resolves to the result of the action.
+    /// </returns>
     [HttpGet("count")]
     [ProducesResponseType<HoursCount>(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status304NotModified)]
     [ProducesResponseType<ValidationProblemDetails>(StatusCodes.Status400BadRequest)]
     [RevalidateCache]
     [Authorize(Policy = "Approved")]
-    public async Task<IActionResult> GetHoursCount(
+    public async Task<ActionResult<HoursCount>> GetHoursCount(
         [FromHeader(Name = "If-None-Match")] string? etag,
         [FromQuery(Name = "date")] DateOnly date,
         [FromQuery(Name = "date-type")] DateType dateType = DateType.Month,
@@ -83,12 +106,19 @@ public class HoursController(IHoursService hoursService, ILogger<HoursController
         return Ok(count);
     }
 
+    /// <summary>
+    /// Accepts a CSV file of hours and processes it.
+    /// </summary>
+    /// <param name="file">The uploaded file data.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the asynchronous operation. Resolves to the result of the action.
+    /// </returns>
     [HttpPost]
-    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(typeof(CountResponse), StatusCodes.Status200OK)]
+    [ProducesResponseType<ProblemDetails>(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType<CountResponse>(StatusCodes.Status200OK)]
     [Authorize(Policy = "Admin")]
     [NotCachedFilter]
-    public async Task<IActionResult> ReceiveHoursFile(IFormFile file)
+    public async Task<ActionResult<CountResponse>> ReceiveHoursFile(IFormFile file)
     {
         using var reader = new StreamReader(file.OpenReadStream());
         using var csv = new CsvReader(reader, CultureInfo.CurrentUICulture);
@@ -116,13 +146,22 @@ public class HoursController(IHoursService hoursService, ILogger<HoursController
         }
     }
 
+    /// <summary>
+    /// Gets the hours trends for a given region.
+    /// </summary>
+    /// <param name="etag">The Etag for the data currently held by the client.</param>
+    /// <param name="region">The region to report for.</param>
+    /// <param name="nhse">Should only NHSE data be included.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the asynchronous operation. Resolves to the result of the action.
+    /// </returns>
     [HttpGet("trends")]
-    [ProducesResponseType(typeof(Trends), StatusCodes.Status200OK)]
+    [ProducesResponseType<Trends>(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status304NotModified)]
     [ProducesResponseType<ValidationProblemDetails>(StatusCodes.Status400BadRequest)]
     [Authorize(Policy = "Lead")]
     [RevalidateCache]
-    public async Task<IActionResult> GetTrends([FromHeader(Name = "If-None-Match")] string? etag, Region region, bool nhse = false)
+    public async Task<ActionResult<Trends>> GetTrends([FromHeader(Name = "If-None-Match")] string? etag, Region region, bool nhse = false)
     {
         if (!Enum.IsDefined(region) || region == Region.Undefined)
         {
