@@ -17,17 +17,28 @@ using System.Security.Claims;
 
 namespace SjaInNumbers.Server.Controllers;
 
+/// <summary>
+/// Controller for managing people data.
+/// </summary>
+/// <param name="personService">The service for managing people data.</param>
 [ApiController]
 [Route("api/people")]
 public class PeopleController(IPersonService personService) : ControllerBase
 {
     private readonly IPersonService personService = personService;
 
+    /// <summary>
+    /// Accepts a CSV file of people data and adds it to the system.
+    /// </summary>
+    /// <param name="file">The uploaded data file.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the asynchronous operation. Resolves to the result of the action.
+    /// </returns>
     [HttpPost]
     [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
     [ProducesResponseType(typeof(CountResponse), StatusCodes.Status200OK)]
     [Authorize(Policy = "Admin")]
-    public async Task<IActionResult> ReceivePersonFile(IFormFile file)
+    public async Task<ActionResult<CountResponse>> ReceivePersonFile(IFormFile file)
     {
         using var reader = new StreamReader(file.OpenReadStream());
         using var csv = new CsvReader(reader, CultureInfo.CurrentUICulture);
@@ -46,12 +57,20 @@ public class PeopleController(IPersonService personService) : ControllerBase
         }
     }
 
+    /// <summary>
+    /// Gets the people activity reports for a specific date and region.
+    /// </summary>
+    /// <param name="etag">The Etag for the data currently held by the client.</param>
+    /// <param name="region">The region to report for.</param>
+    /// <returns>
+    /// A <see cref="Task"/> representing the asynchronous operation. Resolves to the result of the action.
+    /// </returns>
     [HttpGet("reports")]
     [Authorize(Policy = "Lead")]
     [ProducesResponseType(typeof(IAsyncEnumerable<PersonReport>), StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status304NotModified)]
     [ProducesResponseType<ValidationProblemDetails>(StatusCodes.Status400BadRequest)]
-    public async Task<IActionResult> GetReports([FromHeader(Name = "If-None-Match")] string? etag, Region region)
+    public async Task<ActionResult<IAsyncEnumerable<PersonReport>>> GetReports([FromHeader(Name = "If-None-Match")] string? etag, Region region)
     {
         if (!Enum.IsDefined(region) || region == Region.Undefined)
         {
