@@ -7,15 +7,27 @@ using SjaInNumbers.Shared.Model;
 
 namespace SjaInNumbers.Server.Data;
 
+/// <summary>
+/// Contains a set of helper methods for filtering data.
+/// </summary>
 public static class Helpers
 {
+    /// <summary>
+    /// Gets all the active vehicles.
+    /// </summary>
+    /// <param name="vehicles">The vehicles to filter.</param>
+    /// <returns>The filtered items.</returns>
     public static IQueryable<Vehicle> GetActive(this IQueryable<Vehicle> vehicles)
         => vehicles.GetNotDeleted().Where(v => !v.ForDisposal);
 
-    public static IQueryable<T> GetNotDeleted<T>(this IQueryable<T> items)
-        where T : IDeletableItem
-        => items.Where(v => v.Deleted == null);
-
+    /// <summary>
+    /// Gets all the vehicle incidents in a given place.
+    /// </summary>
+    /// <param name="incidents">The vehicle incidents to filter.</param>
+    /// <param name="region">The region to filter for.</param>
+    /// <param name="district">The district to filter for.</param>
+    /// <param name="hub">The hub to filter for.</param>
+    /// <returns>The filtered items.</returns>
     public static IQueryable<VehicleIncident> GetForPlace(this IQueryable<VehicleIncident> incidents, Region region, string? district = null, string? hub = null)
     {
         if (region == Region.All)
@@ -39,6 +51,12 @@ public static class Helpers
         return filteredIncidents;
     }
 
+    /// <summary>
+    /// Gets all the vehicle incidents in a given place.
+    /// </summary>
+    /// <param name="incidents">The vehicle incidents to filter.</param>
+    /// <param name="place">The place to filter for.</param>
+    /// <returns>The filtered items.</returns>
     public static IQueryable<VehicleIncident> GetForPlace(this IQueryable<VehicleIncident> incidents, Place place)
     {
         var district = place.District.Equals("all", StringComparison.OrdinalIgnoreCase) ? null : place.District;
@@ -47,6 +65,12 @@ public static class Helpers
         return GetForPlace(incidents, place.Region, district, hub);
     }
 
+    /// <summary>
+    /// Gets all the vehicles in a given place.
+    /// </summary>
+    /// <param name="vehicles">The vehicles to filter.</param>
+    /// <param name="place">The place to filter for.</param>
+    /// <returns>The filtered items.</returns>
     public static IQueryable<Vehicle> GetForPlace(this IQueryable<Vehicle> vehicles, Place place)
     {
         var district = place.District.Equals("all", StringComparison.OrdinalIgnoreCase) ? null : place.District;
@@ -55,6 +79,14 @@ public static class Helpers
         return GetForPlace(vehicles, place.Region, district, hub);
     }
 
+    /// <summary>
+    /// Gets all the vehicles in a given place.
+    /// </summary>
+    /// <param name="vehicles">The vehicles to filter.</param>
+    /// <param name="region">The region to filter for.</param>
+    /// <param name="district">The district to filter for.</param>
+    /// <param name="hub">The hub to filter for.</param>
+    /// <returns>The filtered items.</returns>
     public static IQueryable<Vehicle> GetForPlace(this IQueryable<Vehicle> vehicles, Region region, string? district = null, string? hub = null)
     {
         if (region == Region.All)
@@ -77,4 +109,14 @@ public static class Helpers
 
         return filteredVehicles;
     }
+
+    /// <summary>
+    /// Gets all the items that have not been deleted.
+    /// </summary>
+    /// <typeparam name="T">The item type to filter by.</typeparam>
+    /// <param name="items">The items to filter.</param>
+    /// <returns>The filtered items.</returns>
+    public static IQueryable<T> GetNotDeleted<T>(this IQueryable<T> items)
+        where T : IDeletableItem
+        => items.Where(v => v.Deleted == null);
 }
